added commands for getting the ign for a discord user and vice versa | set types for discord ids to long because some ids got shortened and couldn't be recognized

+++ src/main/java/de/ximanton/discordverification/commands/VerifyCommand.java
@@ -6,8 +6,6 @@ import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.plugin.Command;
 import de.ximanton.discordverification.DiscordVerification;
 
-import java.math.BigInteger;
-
 public class VerifyCommand extends Command {
 
     public VerifyCommand() {
@@ -29,7 +27,7 @@ public class VerifyCommand extends Command {
             sender.sendMessage(TextComponent.fromLegacyText("this player doesn't exist"));
             return;
         }
-        DiscordVerification.getInstance().getDB().insertPlayer(args[0], BigInteger.valueOf(0), true);
+        DiscordVerification.getInstance().getDB().insertPlayer(args[0], 0L, true);
         sender.sendMessage(TextComponent.fromLegacyText(args[0] + " has been verified"));
     }
 }
